[#195] Don't use PostGIS by default
Consolidate all DB related modules and functions into a single module
for easier encapsulation. The PostGIS database table and the functions
to store PostGIS geometries are no longer called by default.

There is a new config option (`ckan.spatial.use_postgis=true`) to
re-enable this behaviour. This option (and the actual PostGIS modules)
will be dropped in future versions.
