+++ ckanext/spatial/harvested_metadata.py
