--- ckanext/spatial/model/__init__.py
@@ -1,11 +1,0 @@
-from __future__ import absolute_import
-# this is a namespace package
-try:
-    import pkg_resources
-    pkg_resources.declare_namespace(__name__)
-except ImportError:
-    import pkgutil
-    __path__ = pkgutil.extend_path(__path__, __name__)
-
-from .package_extent import *
-from .harvested_metadata import *
