+++ ckanext/spatial/validation/validation.py
@@ -1,6 +1,6 @@
 import os
 from pkg_resources import resource_stream
-from ckanext.spatial.model import ISODocument
+from ckanext.spatial.harvested_metadata import ISODocument
 
 from lxml import etree
